+++ .run/Muster Compliance.run.xml
@@ -1,0 +1,20 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Muster Compliance" type="mocha-javascript-test-runner" nameIsGenerated="true">
+    <node-interpreter>project</node-interpreter>
+    <node-options />
+    <mocha-package>$PROJECT_DIR$/node_modules/mocha</mocha-package>
+    <working-directory>$PROJECT_DIR$/packages/server</working-directory>
+    <pass-parent-env>true</pass-parent-env>
+    <envs>
+      <env name="NO_MOCHA_INTEGRATION" value="true" />
+    </envs>
+    <ui>bdd</ui>
+    <extra-mocha-options>--require ts-node/register/transpile-only -r tsconfig-paths/register</extra-mocha-options>
+    <test-kind>SUITE</test-kind>
+    <test-file>$PROJECT_DIR$/packages/server/src/api/muster/muster.compliance.spec.ts</test-file>
+    <test-names>
+      <name value="Muster Compliance" />
+    </test-names>
+    <method v="2" />
+  </configuration>
+</component>
