+++ packages/server/src/util/test-utils/seed.ts
@@ -1,5 +1,4 @@
 import { getManager } from 'typeorm';
-import { RosterColumnType } from '@covid19-reports/shared';
 import { Observation } from '../../api/observation/observation.model';
 import { Org } from '../../api/org/org.model';
 import { seedOrphanedRecords } from '../../api/orphaned-record/orphaned-record.model.mock';
@@ -7,29 +6,26 @@ import { Role } from '../../api/role/role.model';
 import { User } from '../../api/user/user.model';
 import { Workspace } from '../../api/workspace/workspace.model';
 import { Roster } from '../../api/roster/roster.model';
-import { CustomRosterColumn } from '../../api/roster/custom-roster-column.model';
 import { Unit } from '../../api/unit/unit.model';
 import { env } from '../env';
 import { Log } from '../log';
-import { RosterHistory, ChangeType } from '../../api/roster/roster-history.model';
-import { defaultReportSchemas, ReportSchema } from '../../api/report-schema/report-schema.model';
+import { ChangeType, RosterHistory } from '../../api/roster/roster-history.model';
+import { ReportSchema } from '../../api/report-schema/report-schema.model';
 import { seedOrg } from '../../api/org/org.model.mock';
-import {
-  seedRoleBasicUser,
-  seedRoleAdmin,
-} from '../../api/role/role.model.mock';
+import { seedRoleAdmin, seedRoleBasicUser } from '../../api/role/role.model.mock';
 import { seedUserRole } from '../../api/user/user-role.model.mock';
 import { seedUser } from '../../api/user/user.model.mock';
-import {
-  uniqueEdipi,
-  uniquePhone,
-} from './unique';
+import { uniqueEdipi, uniquePhone } from './unique';
+import { getRosterTestData } from './roster.test-generator';
+import { getCustomRosterColumnTestData } from './custom-roster-column.test-generator';
+import { getUnitsTestData } from './unit.test-generator';
+import { getOrgTestData } from './org-test-generator';
+import { getAdminUserTestData } from './user.test-generator';
+import { getReportSchemaTestData } from './report-schema.test-generator';
 
 require('dotenv').config();
 
 let orgCount = 0;
-let unitCount = 0;
-
 
 export async function seedOrgContactRoles() {
   const { org, contact } = await seedOrgContact();
@@ -63,24 +59,16 @@ export async function seedAll() {
   Log.info('Seeding database...');
 
   // Create Group Admin
-  const groupAdmin = User.create({
-    edipi: uniqueEdipi(),
-    firstName: 'Group',
-    lastName: 'Admin',
-    phone: uniquePhone(),
-    email: `groupadmin@setest.com`,
-    service: 'Space Force',
-    isRegistered: true,
-  });
-  await groupAdmin.save();
+  const adminUser = getAdminUserTestData();
+  await adminUser.save();
 
   const totalAppUsers = 5;
   const totalUnits = 5;
   const totalRosterEntries = 20;
 
   // Create Org 1 & 2 and their Users
-  const orgData1 = await generateOrg(groupAdmin, totalAppUsers, totalRosterEntries, totalUnits);
-  const orgData2 = await generateOrg(groupAdmin, totalAppUsers, totalRosterEntries, totalUnits);
+  const orgData1 = await generateOrg(adminUser, totalAppUsers, totalRosterEntries, totalUnits);
+  const orgData2 = await generateOrg(adminUser, totalAppUsers, totalRosterEntries, totalUnits);
 
   // Set the start date on each roster entry to some time in the past to help with repeatable testing
   await RosterHistory.createQueryBuilder()
@@ -146,53 +134,15 @@ export async function seedAll() {
 
 async function generateOrg(admin: User, numUsers: number, numRosterEntries: number, numUnits: number) {
   orgCount += 1;
-
-  const org = Org.create({
-    name: `Group ${orgCount}`,
-    description: `Group ${orgCount} for testing.`,
-    contact: admin,
-    indexPrefix: `testgroup${orgCount}`,
-    reportingGroup: `test${orgCount}`,
-    defaultMusterConfiguration: [],
-  });
+  const org = getOrgTestData(admin, orgCount);
   await org.save();
-
-  let reportSchemas = ReportSchema.create(defaultReportSchemas);
-  reportSchemas = reportSchemas.concat(ReportSchema.create(defaultReportSchemas));
-  for (let i = 0; i < reportSchemas.length; i++) {
-    const report = reportSchemas[i];
-    report.org = org;
-    if (i > 0) {
-      report.id += i.toString();
-    }
-  }
+  const reportSchemas = getReportSchemaTestData(org);
   await ReportSchema.save(reportSchemas);
-
-  const customColumn = CustomRosterColumn.create({
-    org,
-    display: `My Custom Column ${orgCount}`,
-    type: RosterColumnType.String,
-    phi: false,
-    pii: false,
-    required: false,
-  });
+  const customColumn = getCustomRosterColumnTestData(org, orgCount);
   await customColumn.save();
 
   // Add Units
-  const units: Unit[] = [];
-  for (let i = 1; i <= numUnits; i++) {
-    unitCount += 1;
-    const unit = Unit.create({
-      org,
-      name: `Unit ${unitCount}`,
-      musterConfiguration: [
-        { days: 62, startTime: '00:00', timezone: 'America/Los_Angeles', durationMinutes: 120, reportId: reportSchemas[0].id },
-        { startTime: '2020-01-02T02:00:00.000', timezone: 'America/Los_Angeles', durationMinutes: 120, reportId: reportSchemas[1].id },
-      ],
-      includeDefaultConfig: true,
-    });
-    units.push(unit);
-  }
+  const units = getUnitsTestData(numUnits, org, reportSchemas);
   await Unit.save(units);
 
   const groupAdminRole = createGroupAdminRole(org);
@@ -222,21 +172,7 @@ async function generateOrg(admin: User, numUsers: number, numRosterEntries: numb
   }
 
   // Add roster entries
-  const rosterEntries: Roster[] = [];
-  for (let i = 0; i < numRosterEntries; i++) {
-    const customColumns: any = {};
-    customColumns[customColumn.name] = `custom column value`;
-    const edipi = uniqueEdipi();
-    const edipiNum = parseInt(edipi);
-    const rosterEntry = Roster.create({
-      edipi,
-      firstName: `RosterFirst${edipiNum}`,
-      lastName: `RosterLast${edipiNum}`,
-      unit: units[i % numUnits],
-      customColumns,
-    });
-    rosterEntries.push(rosterEntry);
-  }
+  const rosterEntries = getRosterTestData(numRosterEntries, customColumn, units, numUnits);
   await Roster.save(rosterEntries);
 
   // Insert observations per roster entry
