+++ packages/server/src/util/test-utils/user.test-generator.ts
@@ -1,0 +1,14 @@
+import { User } from '../../api/user/user.model';
+import { uniqueEdipi, uniquePhone } from './unique';
+
+export function getAdminUserTestData() {
+  return User.create({
+    edipi: uniqueEdipi(),
+    firstName: 'Group',
+    lastName: 'Admin',
+    phone: uniquePhone(),
+    email: `groupadmin@setest.com`,
+    service: 'Space Force',
+    isRegistered: true,
+  });
+}
